+++ eslint.config.js
@@ -1,0 +1,25 @@
+import js from "@eslint/js";
+import eslintPluginPrettierRecommended from "eslint-plugin-prettier/recommended";
+import { defineConfig, globalIgnores } from "eslint/config";
+import globals from "globals";
+import tseslint from "typescript-eslint";
+
+export default defineConfig([
+    globalIgnores(["dist/*"]),
+    {
+        files: ["**/*.{js,mjs,cjs,ts,mts,cts}"],
+        plugins: {
+            js
+        },
+        extends: ["js/recommended"],
+        languageOptions: { globals: globals.browser },
+        rules: {
+            "@typescript-eslint/no-unused-vars": [
+                "error",
+                { argsIgnorePattern: "^_", varsIgnorePattern: "^_", caughtErrorsIgnorePattern: "^_" }
+            ]
+        }
+    },
+    tseslint.configs.recommended,
+    eslintPluginPrettierRecommended
+]);