+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [7.2.0] - 2026-02-03
+
+### Added
+
+- Support for Dynamic Tokens
+    - PF2e Effects Halo will now "calculate" the size of the token and determine how many effect icons it can display around the token. This applies to every "row" of effects around the token meaning that more effects can be displayed on the 2nd "row" and even more on the 3rd "row".
+- PF2e/SF2e Effect Colors
+    - By default the background/border of the effects will depend on which system you are using; however, this is configurable per user in the settings if you want to change the colors.
+- Option to increase the spacing of effects on the same row
+- Option to increase the spacing between rows
+- Option to increase the "scale of the halo
+    - What this setting does is increase the calculated radius of the tokens, so if you are using funky tokens that need more or less space between the token and the first row of effects you can use this setting.
+    - There is a "global" setting in the settings along with a setting stored on each actor, the actor settings can be accessed by opening the actor sheet you want to modify and clicking on the "magic sparkles" icon in the title bar. Please note that actor settings are stored on the actor itself, so if you modify a non-linked actor the settings are stored on the token itself and won't apply to other tokens of the same actor. You would need to modify the parent actor first and make tokens with it to apply the settings to multiple non-linked tokens.
+
 ## [7.1.0] - 2026-02-01
 
 ### Added
@@ -34,7 +48,8 @@ Foundry VTT Version 13 Support
 - Removed PF2e Dorako UX, this module is always on when enabled
 - Removed PF2e Dorako UI theme check, this module uses the default
 
-[Unreleased]: https://github.com/7H3LaughingMan/pf2e-effects-halo/compare/v7.1.0...HEAD
+[Unreleased]: https://github.com/7H3LaughingMan/pf2e-effects-halo/compare/v7.2.0...HEAD
+[7.2.0]: https://github.com/7H3LaughingMan/pf2e-effects-halo/compare/v7.1.0...v7.2.0
 [7.1.0]: https://github.com/7H3LaughingMan/pf2e-effects-halo/compare/v7.0.0...v7.1.0
 [7.0.0]: https://github.com/7H3LaughingMan/pf2e-effects-halo/compare/v6.1.0...v7.0.0
 [6.1.0]: https://github.com/7H3LaughingMan/pf2e-effects-halo/compare/v6.0.0...v6.1.0